+++ splunk-otel-android/src/main/java/com/splunk/rum/SplunkRum.java
@@ -27,7 +27,6 @@ import android.os.Looper;
 import android.util.Log;
 import android.webkit.WebView;
 import androidx.annotation.Nullable;
-import com.splunk.rum.internal.opentelemetry.GlobalAttributesSpanAppender;
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
@@ -36,6 +35,7 @@ import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.StatusCode;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.instrumentation.okhttp.v3_0.OkHttpTelemetry;
+import io.opentelemetry.rum.internal.GlobalAttributesSpanAppender;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
 import java.util.concurrent.TimeUnit;
