+++ splunk-otel-android/src/test/java/com/splunk/rum/SplunkRumTest.java
@@ -36,13 +36,13 @@ import android.app.Application;
 import android.content.Context;
 import android.location.Location;
 import android.webkit.WebView;
-import com.splunk.rum.internal.opentelemetry.GlobalAttributesSpanAppender;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanId;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.context.Scope;
+import io.opentelemetry.rum.internal.GlobalAttributesSpanAppender;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.sdk.testing.exporter.InMemorySpanExporter;
 import io.opentelemetry.sdk.testing.junit4.OpenTelemetryRule;
